Pushed Settings component down by one level to faciliate filtering/search
